inclusion of about us content/possible footer
In this Git commit I am committing the initial work that I have done for the about us page and a possible footer for both the HTML structure and the stylings. I tried to make everything for the page as responsive as it could be so the text is visible for different screen widths, so hopefully it is fine.