Cleaning Up CSS and HTML
Cleaned up the CSS and HTML for the about us page. Put options in the navbar for the about us page and goto quiz test page (will probably take this out), and cleaned up a lot of the media calls since with having the content be scrollable it doesn't seem as necessary.