Worked on github OAuth login implementation
Created the basic server/client behavior and structure for the GitHub OAuth login.  This includes some reconfiguring of the index.html navbar for the categories page as well as the behavior in server.js, the client behavior in index.js, and some package adjustments to include cors so that we get no CORS errors.