Some quick fixes on the quiz html and about us css
Fixed the quiz buttons so that now clicking anywhere on the button will do the correct actions for the quiz, rather than before where some areas would not carry the category over since the link was encased in the button. Also ended up removing a line of still referencing the footer for the about us page.