Added jshint and commands to package
Added the JSHint installation into the packages using "npm install jshint", and then from there also added the ability to run JSHint for all of the current JavaScript files that we have. Mainly adds more functionality in terms of the build and error checking to get any recommendations for the JavaScript.